Add files via upload
Exploratory Data Analysis (EDA) is a critical step in the data analysis process. It involves visually and statistically summarizing, interpreting, and making initial observations about a dataset. Python offers a variety of libraries and tools that are commonly used for EDA, including:

Pandas: Pandas is a popular library for data manipulation and analysis. It provides data structures like DataFrames that are particularly useful for cleaning and exploring data.

Matplotlib: Matplotlib is a widely used 2D plotting library that provides a flexible interface for creating a wide range of visualizations.

Seaborn: Seaborn is built on top of Matplotlib and provides a higher-level interface for creating attractive and informative statistical graphics.

Plotly: Plotly is an interactive graphing library that enables you to create interactive, web-based visualizations.

NumPy: NumPy is the fundamental package for scientific computing in Python. It provides support for working with arrays and mathematical functions.

Statistical Functions: Python's built-in scipy.stats module provides various statistical functions and distributions that are useful for exploring data.